+++ src/components/CommandActionIcon.tsx
@@ -1,11 +1,11 @@
-import { ActionIcon, ActionIconProps } from "@mantine/core"
+import { ActionIcon, ActionIconProps, Tooltip } from "@mantine/core"
 import { executeCommandSequence } from "../commands"
 import { setLocalStorage } from "../utilities/hooks"
 
 const CommandActionIcon = (
-  props: ActionIconProps & { component?: any; command: string },
+  props: ActionIconProps & { component?: any; command: string; label?: string },
 ) => {
-  return (
+  const icon = (
     <ActionIcon
       onMouseEnter={() => setLocalStorage("Hover Command", props.command)}
       onMouseLeave={() => setLocalStorage("Hover Command", null)}
@@ -17,6 +17,12 @@ const CommandActionIcon = (
       {...props}
     />
   )
+
+  if (props.label) {
+    return <Tooltip label={props.label}>{icon}</Tooltip>
+  }
+
+  return icon
 }
 
 export default CommandActionIcon