+++ src/components/plugins/CurrentEditorPlugin.tsx
@@ -1,4 +1,5 @@
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext"
+import { COMMAND_PRIORITY_HIGH, FOCUS_COMMAND } from "lexical"
 import { useEffect } from "react"
 
 const CurrentEditorPlugin = () => {
@@ -12,6 +13,18 @@ const CurrentEditorPlugin = () => {
         window.currentEditor = undefined
       }
     }
+  }, [editor])
+
+  useEffect(() => {
+    return editor.registerCommand(
+      FOCUS_COMMAND,
+      () => {
+        window.currentEditor = editor
+
+        return false
+      },
+      COMMAND_PRIORITY_HIGH,
+    )
   }, [editor])
 
   return null